Rename SIPPY to sippy to match python conventions
