+++ setup.py
@@ -4,13 +4,13 @@ with open('README.md', 'r') as f:
 	readme = f.read()
 
 setuptools.setup(
-	name="SIPPY",
+	name="sippy",
 	version="0.1.0",
 	author="Giuseppe Armenise",
 	description="Systems Identification Package for Python",
 	long_description=readme,
 	long_description_content_type='text/markdown',
-	url="https://github.com/CPCLAB-UNIPI/SIPPY",
+	url="https://github.com/CPCLAB-UNIPI/sippy",
 	packages=setuptools.find_packages(),
 	python_requires="==2.7,>=3.5,<=3.6",
 	install_requires=(
